refactor: 重构项目结构并迁移文件至新目录
重构项目结构，将组件、类型定义、工具函数和服务分离到各自的目录中。主要变更包括：
1. 将 `src/app/components` 拆分为 `ui`、`layout` 和 `common` 子目录
2. 创建 `src/types` 存放类型定义，从原 model.tsx 文件中提取
3. 创建 `src/services` 存放 API 请求逻辑
4. 创建 `src/utils` 存放工具函数
5. 更新所有相关文件的引用路径
6. 修复因重构导致的类型错误和导入问题
