+++ src/app/music/components/SongItemSkeleton.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from "@/app/components/Skeleton"
+import { Skeleton } from "@/components/ui/Skeleton"
 
 export default function SongItemSkeleton() {
   return (
