+++ src/app/music/page.tsx
@@ -1,12 +1,13 @@
 "use client"
 
 import { useState, useEffect, useCallback, useRef } from "react"
-import { Song, transferVersion } from "@/app/music/songModel"
-import SongList from "@/app/music/songList"
-// import LoadingSpinner from "../components/LoadingSpinner"
-import ActionButton from "../components/ActionButton"
+import { Song } from "@/types/music"
+import { transferVersion } from "@/utils/music"
+import SongList from "./songList"
+// import LoadingSpinner from "@/components/ui/LoadingSpinner"
+import ActionButton from "@/components/ui/ActionButton"
 import type { Step } from "react-joyride"
-import Guide from "../components/Guide"
+import Guide from "@/components/common/Guide"
 // import { FaFilter, FaTimes } from "react-icons/fa"
 import { AnimatePresence, motion } from "framer-motion"
 // import Notice from "../components/Notice"
