+++ src/app/tool/best/page.tsx
@@ -1,10 +1,10 @@
 "use client"
 
-import Button from "@/app/components/button/Button"
-import Guide from "@/app/components/Guide"
-import LoadingSpinner from "@/app/components/LoadingSpinner"
-import { ShareableImageSub } from "@/app/components/ShareableImage"
-import { UserProfile } from "@/app/user/model"
+import Button from "@/components/ui/button/Button"
+import Guide from "@/components/common/Guide"
+import LoadingSpinner from "@/components/ui/LoadingSpinner"
+import { ShareableImageSub } from "@/components/common/ShareableImage"
+import { UserProfile } from "@/types/user"
 import { useCallback, useEffect, useState } from "react"
 import { Step } from "react-joyride"
 import { CONFIG } from "@/config/api"
@@ -332,7 +332,7 @@ export default function BestPage() {
     }
   }, [nowFrom, accounts, token])
   useEffect(() => {
-    // const data = { "rating": 15468, "rating_b35": 10833, "rating_b15": 4635, "scores_b35": [{ "id": 1343, "title": "マツヨイナイトバグ", "level": "13+", "level_index": 3, "achievements": 100.5263, "fc": 3, "fs": 0, "dx_score": 2747, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 400, "title": "デッドレッドガールズ", "level": "13+", "level_index": 3, "achievements": 100.6019, "fc": null, "fs": 0, "dx_score": 2494, "dx_rating": 312.0, "rate": 0, "type": "standard" }, { "id": 1096, "title": "モ°ルモ°ル", "level": "13+", "level_index": 3, "achievements": 100.6086, "fc": 3, "fs": 0, "dx_score": 2487, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 1236, "title": "Last Samurai", "level": "13+", "level_index": 3, "achievements": 100.5609, "fc": 3, "fs": 2, "dx_score": 1200, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 1466, "title": "群青シグナル", "level": "13+", "level_index": 3, "achievements": 100.5721, "fc": null, "fs": 0, "dx_score": 2661, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 1310, "title": "Trick tear", "level": "14", "level_index": 3, "achievements": 100.2449, "fc": null, "fs": 0, "dx_score": 2246, "dx_rating": 311.0, "rate": 1, "type": "dx" }, { "id": 1461, "title": "#狂った民族２ PRAVARGYAZOOQA", "level": "14", "level_index": 3, "achievements": 100.1856, "fc": null, "fs": 0, "dx_score": 2513, "dx_rating": 311.0, "rate": 1, "type": "dx" }, { "id": 1573, "title": "Final Step!", "level": "13+", "level_index": 3, "achievements": 100.7084, "fc": null, "fs": 0, "dx_score": 2347, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1566, "title": "Knight Rider", "level": "13+", "level_index": 3, "achievements": 100.772, "fc": null, "fs": 0, "dx_score": 2414, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 701, "title": "Doll Judgment", "level": "13+", "level_index": 3, "achievements": 100.5674, "fc": null, "fs": 0, "dx_score": 2389, "dx_rating": 310.0, "rate": 0, "type": "standard" }, { "id": 1288, "title": "Big Daddy", "level": "13+", "level_index": 3, "achievements": 100.5366, "fc": null, "fs": 0, "dx_score": 2705, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 841, "title": "終点", "level": "13+", "level_index": 3, "achievements": 100.7131, "fc": 3, "fs": 0, "dx_score": 1973, "dx_rating": 310.0, "rate": 0, "type": "standard" }, { "id": 711, "title": "拝啓ドッペルゲンガー", "level": "13+", "level_index": 3, "achievements": 100.506, "fc": null, "fs": 0, "dx_score": 2815, "dx_rating": 310.0, "rate": 0, "type": "standard" }, { "id": 288, "title": "六兆年と一夜物語", "level": "13+", "level_index": 3, "achievements": 100.8854, "fc": 3, "fs": 0, "dx_score": 2037, "dx_rating": 310.0, "rate": 0, "type": "standard" }, { "id": 1453, "title": "Rainbow Rush Story", "level": "13+", "level_index": 3, "achievements": 100.7663, "fc": null, "fs": 0, "dx_score": 2514, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1459, "title": "You Mean the World to Me", "level": "13+", "level_index": 3, "achievements": 100.5134, "fc": null, "fs": 0, "dx_score": 1879, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 548, "title": "DETARAME ROCK&ROLL THEORY", "level": "13+", "level_index": 3, "achievements": 100.686, "fc": 3, "fs": 0, "dx_score": 2191, "dx_rating": 310.0, "rate": 0, "type": "standard" }, { "id": 1524, "title": "Alice's Suitcase", "level": "13+", "level_index": 3, "achievements": 100.6736, "fc": 3, "fs": 0, "dx_score": 2144, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1576, "title": "Cthugha", "level": "14", "level_index": 3, "achievements": 100.139, "fc": null, "fs": 0, "dx_score": 3095, "dx_rating": 309.0, "rate": 1, "type": "dx" }, { "id": 1176, "title": "Climax", "level": "14", "level_index": 3, "achievements": 100.2156, "fc": null, "fs": 0, "dx_score": 3017, "dx_rating": 309.0, "rate": 1, "type": "dx" }, { "id": 1475, "title": "SUPER AMBULANCE", "level": "14", "level_index": 3, "achievements": 100.0727, "fc": null, "fs": 0, "dx_score": 2760, "dx_rating": 309.0, "rate": 1, "type": "dx" }, { "id": 379, "title": "Caliburne ～Story of the Legendary sword～", "level": "14", "level_index": 3, "achievements": 100.2489, "fc": null, "fs": 0, "dx_score": 2517, "dx_rating": 309.0, "rate": 1, "type": "standard" }, { "id": 1479, "title": "Hainuwele", "level": "14", "level_index": 3, "achievements": 100.0418, "fc": null, "fs": 0, "dx_score": 2685, "dx_rating": 309.0, "rate": 1, "type": "dx" }, { "id": 1231, "title": "生命不詳", "level": "13+", "level_index": 3, "achievements": 100.8213, "fc": 3, "fs": 0, "dx_score": 2329, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1022, "title": "TwisteD! XD", "level": "13+", "level_index": 3, "achievements": 100.5343, "fc": null, "fs": 0, "dx_score": 2414, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1296, "title": "とびだせ！TO THE COSMIC!!", "level": "13+", "level_index": 3, "achievements": 100.724, "fc": 2, "fs": 0, "dx_score": 2564, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1447, "title": "エゴロック", "level": "13+", "level_index": 3, "achievements": 100.6095, "fc": null, "fs": 0, "dx_score": 1928, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 461, "title": "無敵We are one!!", "level": "13+", "level_index": 3, "achievements": 100.6059, "fc": 3, "fs": 0, "dx_score": 2525, "dx_rating": 308.0, "rate": 0, "type": "standard" }, { "id": 1143, "title": "アトロポスと最果の探究者", "level": "13+", "level_index": 3, "achievements": 100.6573, "fc": null, "fs": 0, "dx_score": 2194, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1445, "title": "遺伝子レベル∞スパイラル", "level": "13+", "level_index": 3, "achievements": 100.6425, "fc": 3, "fs": 0, "dx_score": 2263, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1159, "title": "Beautiful Future", "level": "13+", "level_index": 3, "achievements": 100.5278, "fc": 3, "fs": 0, "dx_score": 2236, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1195, "title": "マネマネサイコトロピック", "level": "13+", "level_index": 3, "achievements": 100.5842, "fc": 3, "fs": 2, "dx_score": 2290, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 561, "title": "いっしそう電☆舞舞神拳！", "level": "13+", "level_index": 3, "achievements": 100.5671, "fc": 3, "fs": 0, "dx_score": 2353, "dx_rating": 308.0, "rate": 0, "type": "standard" }, { "id": 1208, "title": "Cyaegha", "level": "13+", "level_index": 3, "achievements": 100.505, "fc": 3, "fs": 0, "dx_score": 2543, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 849, "title": "Kattobi KEIKYU Rider", "level": "13+", "level_index": 3, "achievements": 100.5978, "fc": 3, "fs": 2, "dx_score": 2319, "dx_rating": 308.0, "rate": 0, "type": "standard" }], "scores_b15": [{ "id": 1527, "title": "enchanted wanderer", "level": "13+", "level_index": 3, "achievements": 100.7009, "fc": 2, "fs": 0, "dx_score": 1653, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1588, "title": "Complex Mind", "level": "13+", "level_index": 3, "achievements": 100.6411, "fc": null, "fs": 0, "dx_score": 2357, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1587, "title": "Halfway(>∀<)", "level": "13+", "level_index": 3, "achievements": 100.6558, "fc": 3, "fs": 0, "dx_score": 2476, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1542, "title": "ここからはじまるプロローグ。 (Kanon Remix)", "level": "13+", "level_index": 3, "achievements": 100.578, "fc": null, "fs": 0, "dx_score": 2536, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1553, "title": "Last Kingdom", "level": "13+", "level_index": 3, "achievements": 100.7854, "fc": 2, "fs": 0, "dx_score": 2328, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1590, "title": "あつすぎの歌", "level": "13+", "level_index": 3, "achievements": 100.7567, "fc": 3, "fs": 2, "dx_score": 2217, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1633, "title": "OMAKENO Stroke", "level": "13+", "level_index": 3, "achievements": 100.8466, "fc": 2, "fs": 0, "dx_score": 2033, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1546, "title": "地球", "level": "14", "level_index": 3, "achievements": 100.2835, "fc": null, "fs": 0, "dx_score": 2351, "dx_rating": 309.0, "rate": 1, "type": "dx" }, { "id": 1561, "title": "おとせサンダー", "level": "13+", "level_index": 4, "achievements": 100.5527, "fc": null, "fs": 0, "dx_score": 2621, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1556, "title": "Hello, Hologram", "level": "13+", "level_index": 3, "achievements": 100.7926, "fc": 2, "fs": 0, "dx_score": 2430, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1539, "title": "リフヴェイン", "level": "13+", "level_index": 3, "achievements": 100.5441, "fc": 3, "fs": 2, "dx_score": 1772, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1584, "title": "コンティニュー！ feat. 藍月なくる", "level": "13+", "level_index": 3, "achievements": 100.6865, "fc": 3, "fs": 0, "dx_score": 2434, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1543, "title": "モ°ルモ°ル (MZK Skippin' Remix)", "level": "13+", "level_index": 3, "achievements": 100.6718, "fc": 2, "fs": 2, "dx_score": 1885, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1533, "title": "にゃーにゃー冒険譚", "level": "13+", "level_index": 3, "achievements": 100.5203, "fc": null, "fs": 0, "dx_score": 2934, "dx_rating": 308.0, "rate": 0, "type": "dx" }, { "id": 1604, "title": "『ウソテイ』 ～一回戦せりなvsしろなvsなずな～", "level": "13+", "level_index": 3, "achievements": 100.5811, "fc": 3, "fs": 0, "dx_score": 2406, "dx_rating": 308.0, "rate": 0, "type": "dx" }] }
+    // const data = { "rating": 15468, "rating_b35": 10833, "rating_b15": 4635, "scores_b35": [{ "id": 1343, "title": "マツヨイナイトバグ", "level": "13+", "level_index": 3, "achievements": 100.5263, "fc": 3, "fs": 0, "dx_score": 2747, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 400, "title": "デッドレッドガールズ", "level": "13+", "level_index": 3, "achievements": 100.6019, "fc": null, "fs": 0, "dx_score": 2494, "dx_rating": 312.0, "rate": 0, "type": "standard" }, { "id": 1096, "title": "モ°ルモ°ル", "level": "13+", "level_index": 3, "achievements": 100.6086, "fc": 3, "fs": 0, "dx_score": 2487, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 1236, "title": "Last Samurai", "level": "13+", "level_index": 3, "achievements": 100.5609, "fc": 3, "fs": 2, "dx_score": 1200, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 1466, "title": "群青シグナル", "level": "13+", "level_index": 3, "achievements": 100.5721, "fc": null, "fs": 0, "dx_score": 2661, "dx_rating": 312.0, "rate": 0, "type": "dx" }, { "id": 1310, "title": "Trick tear", "level": "14", "level_index": 3, "achievements": 100.2449, "fc": null, "fs": 0, "dx_score": 2246, "dx_rating": 311.0, "rate": 1, "type": "dx" }, { "id": 1461, "title": "#狂った民族２ PRAVARGYAZOOQA", "level": "14", "level_index": 3, "achievements": 100.1856, "fc": null, "fs": 0, "dx_score": 2513, "dx_rating": 311.0, "rate": 1, "type": "dx" }, { "id": 1573, "title": "Final Step!", "level": "13+", "level_index": 3, "achievements": 100.7084, "fc": null, "fs": 0, "dx_score": 2347, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 1566, "title": "Knight Rider", "level": "13+", "level_index": 3, "achievements": 100.772, "fc": null, "fs": 0, "dx_score": 2414, "dx_rating": 310.0, "rate": 0, "type": "dx" }, { "id": 701, "title": "Doll Judgment", "level": "13+", "level_index": 3, "achievements": 100.5674, "fc": null, "fs": 0, "dx_score": 2389, "dx_rating": 310.0, "rate": 0, "type": "standard" }, { "id": 1288, "title": "Big Daddy", "level": "13+", "level_index":[... 7377 chars omitted ...]
     // setBest35(data.scores_b35)
     // setBest15(data.scores_b15)
     // setRating15(data.rating_b15)
@@ -505,7 +505,7 @@ export default function BestPage() {
           </div>
         </div>
 
-        <div className="w-full flex flex-col items-center">
+        <div className="w-full flex flex-col items-center z-[10]">
           <div id="b50" className="w-full max-w-[1600px] px-2">
             <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-5 gap-3 w-full justify-items-center">
               {isLoading ? (
