+++ src/components/common/Guide.tsx
