+++ src/components/common/NewsCard.tsx
