+++ src/components/common/SearchGameCenter.tsx
