+++ src/components/common/SearchGameCenterResults.tsx
