+++ src/components/ui/AnimatedComponent.tsx
