+++ src/components/ui/SvgStrokedText.tsx
